feat: recursive folding and decider verifier for Protogalaxy (#4156)
Implements the recursive folding verifier and recursive decider verifier
resembling the native ones as closely as possible. Tests follow the
example of the other recursive honk verifiers and native protogalaxy
tests.

Additionally, I introduced a function that sequentially computes
polynomials in coefficient form in the `ProtoGalaxyRecursiveVerifier` to
avoid having to instantiate the entire `Polynomial` class on
`stdlib::bn254`.

Closes https://github.com/AztecProtocol/barretenberg/issues/789.
Closes https://github.com/AztecProtocol/barretenberg/issues/834.