feat: implement bigint in Noir, using bigint opcodes (#4198)
This PR enables big integers in Noir using the bigint opcodes.
It does not implement bigints in ACVM which is required to solve the
opcodes.
It also does not implement bigints opcodes in Barretenberg, which is
required to generate the constraints.

However it provides a BigInt structure in the stdlib so you can write
code like this in Noir, which will be compiled into ACIR using the
bigint opcodes.

```
    let a = bigint::BigInt::from_le_bytes([3], [23]);
    let b = bigint::BigInt::from_le_bytes([5], [23]);
    let c = a+b;
    let bytes = c.to_le_bytes();
    std::println(bytes[0]);

```

I did not add a test yet because ACMV solver is not yet implemented.

---------

Co-authored-by: kevaundray <kevtheappdev@gmail.com>