feat: Contract classes and instances (#4192)
Adds `ContractInstance` and `ContractClass` interfaces as specified in
the yellow paper, with the addition of `PublicFunctions` in
`ContractClasses` until we remove them as first class citizens from the
protocol. Also includes hashes for contract classes and artifacts.
Stores contract classes and instances in the node in new tables, as well
as artifacts and instances in the pxe.

Fixes #4053