feat!: add opcode for sha256 compression function (#4229)
The opcode will allow SHA256 implementation in Noir with support to
variable length input, while keeping BB native performance.

The PR includes usage of the function in stdlib, but does not include
implementation of the function. It should be added in a future PR for
ACVM, BB, Brillig and Noir (for constant folding).