chore: delete MemoryDB (#4241)
This module is not being used anywhere and was just being kept around as
a reference. However this results in a maintenance burden as this shares
tests with the KV DB, which means we have to keep feature parity (see
e.g. https://github.com/AztecProtocol/aztec-packages/pull/4238).

The KV DB is in a good place already, and we can always look at the repo
history if we wish to bring this back, so the simplest path forward is
to remove this.