feat: add poseidon2 hashing to native transcript (#3718)
Change the transcript proof to be a vector of field elements instead of
a vector of bytes. We want this more crucially for the stdlib
transcript, but the native transcript should be implemented similarly.

The key contribution is the conversion functions in
field_conversion.hpp/cpp, which allow for types, in particular
grumpkin::fr, to be converted to and from bb::fr elements.

Some notes about why this was chosen:
https://hackmd.io/2XU21v9WTx6XUSnMgZIUig. In short, it is the most
efficient in terms of the number of gates.

Updates the transcript deserialization/serialization and corresponding
tests and makes other style updates.

Resolves https://github.com/AztecProtocol/barretenberg/issues/777

Also some notes from trying to resolve bb.js ACIR tests:
https://hackmd.io/OrQUD_nTQLeIZVvy4zpzHA