refactor: `toFields()`/`fromFields(...)` methods in more classes (#4335)
Partially fixes #4332

The goal of this PR was to progress on nuking
yarn-project/acir-simulator/src/acvm/serialize.ts and move it to the
individual classes to then be able to easily test it in the classes
tests and improve readability.

I didn't fully tackle the issue in this one because it was all getting
too big.