refactor: eth address tech debt cleanup (#4442)
We used to serialize EthAddress as 32 bytes because we needed that for
compatibility with CBINDs. Given that we no longer use the C++ code this
is no longer necessary.

**Note**: decided to not completely nuke `EthAddress.toBuffer32()`
because that would require updating how we compute calldata/body hash
and it makes sense to do that once we'll do that in #3938 (linked it in
the issue).