feat: add hashing to stdlib transcript (#4161)
Combines native and stdlib transcripts by templating it with a
TranscriptParams parameter. This parameter contains the type of field
element in the proof (bb::fr for native and field_t<Builder> for stdlib)
and conversion functions for converting from the field element to/from
other types.

Adds the stdlib conversion functions, and the convert_challenge function
for native and stdlib for turning a challenge into a bn254 scalar or
grumpkin scalar.

Resolves https://github.com/AztecProtocol/barretenberg/issues/835.

---------

Co-authored-by: codygunton <codygunton@gmail.com>