feat: add fee payment methods (#4504)
This PR updates the account entrypoints to add support for fee payments.
The concept of a fee payment method has been added, with two examples: a
native token payment method (where the caller pays in the native gas
token) and a generic payment method (where the sender can use a fee
preparation contract).