chore: updating viem (#4783)
`e2e-public-cross-chain-messaging` is flaky and it's caused by a failure
to retrieve logs from anvil. The low hanging fruit here was to try to
update viem and see whether by any chance this gets resolved. I don't
think viem is the culprit but given that the version we used now was
ancient, they fixed a lot of bugs and we'll need to do the update anyway
to get blobs support it made sense to just do it.