feat: Parallel native/wasm bb builds. Better messaging around using ci cache. (#4766)
Ignore badly named branch name. Public ECR is a bit of a no go as it has
a bunch of restrictions that is a problem:
* Only us-east-1.
* Missing various API's, like ability to retag remote images without
pulling.
* No lifecycle policies.

It was actually the second point that made me give up on public ecr
right now.

* Once upon a time, we published our deployment containers to a private
ECR repo in eu-west-2. Removes all references to the old private ECR
repo from dockerfiles and build stuff, now just using the
`aztecprotocol` dockerhub org.
* Makes `./bootstrap.sh` require an explicit command, one of
`full|fast|clean`. Add warnings when trying to use cache but can't, or
not using cache but can. By being explicit users are made aware of the
option of a `fast` bootstrap.
* barretenberg now builds native/wasm/wasm-threads together in parallel,
with pretty line coloring to distinguish output.
* New amis for build-instances (Ubuntu 22 with upgraded aws-cli).