feat: Macros for initializer checks (#4830)
Adds two new macros for private functions `#[aztec(initializer)]` and
`#[aztec(initcheck)]` that add the `mark_as_initialized` and
`assert_is_initialized` calls respectively. The idea is to eventually
drop the `initcheck` and make it the default, and add a `noinitcheck`
for the cases where we don't want initializer checks.

Builds on #4807