feat(avm): integrate AVM with initializers (#5469)
This PR creates private/public/avm variants of some lib functions used for initializers. This is because in general the flow in private/public/avm will be substantially different. The public version will be removed once the AVM is fully enabled.

I'm adding a new AVM test contract because: having an initializer adds a check to EVERY other function in the contract. This makes the unit tests fail.