feat: /foundry is canoncial build of foundry. e2e tests can start own anvil. (#5522)
We want e2e tests to be able to launch their own anvil. This will enable
deterministic L1 states for the test.
* We use viems anvil package and get-port to find a random port to run
anvil on, and launch it and use it if no `ETHEREUM_HOST` env var is set.
* Anvil is wrapped in wrapper script to ensure it's process is killed
when the parent dies, no matter what.
* This is as close as we can get to what looks like an "in-memory"
blockchain for testing with no side-effects.
* Anvil needs to be present *within* the e2e image to achieve this. This
means we need an x86 and arm build of anvil in our image, and we want it
to be whatever our chosen version of foundry is (which is often some
random pinned nightly that only has x86 builds).
* Solution, foundry build in monorepo! `/foundry` now contains an
Earthly ubuntu build of foundry using their `local` build target which
keeps builds at about 1m on a decent core machine.
* L1 contracts, yarn-project, and any other code that uses foundry,
should take foundry from the foundry image we build.
* We also do some cleanup in `yarn-project/Earthfile`, and leverage
layer caching to only reinstall dependencies if bb.js, noir-packages, or
package.json files change.

---------

Co-authored-by: ludamad <domuradical@gmail.com>
Co-authored-by: ludamad <adam.domurad@gmail.com>