feat: Parsing non-string assertion payloads in noir js (#6079)
This PR adds a noirc_abi_wasm entrypoint for decoding ABI error types.
This entrypoint is used by noir_js when an execution error is detected
to try to decode the error using the ABI.
If the result from decoding the error is a string (such as a fmt string
assertion, that previously weren't decoded at all) it'll patch the error
message to include the formatted string.
If the result from decoding the error is a non-string payload (structs,
tuples, arrays, etc) it'll add the decodedPayload property to the error
so users can get the data back from the circuit.
Also, this PR refactors ErrorSelector so it's a shared struct, instead
of being casted to u64 outside of the compiler backend.