feat: prototype for using the databus with ACIR opcode (#6366)
Related to https://github.com/noir-lang/noir/issues/4974

The PR implements logic in BB to handle ACIR opcodes for the databus.
Because the databus in BB works only with Goblin which is not yet
supported by Noir, the code is not fully tested, so this is only
prototype code.
Using the databus with UltraPlonk will result in a failing ASSERT.
I have updated the BB unit test to work with GoblinPlonk, so the BB code
is tested.

---------

Co-authored-by: ludamad <adam.domurad@gmail.com>
Co-authored-by: ledwards2225 <l.edwards.d@gmail.com>