chore: dedup public data writes in kernel (#6749)
We now dedup public data writes in the public kernel tail. We do this by
first sorting them in TS based FIRST on leaf index, and next on side
effect counter, and get hints for that reordering (which we assert in
the kernel).

We then deduplicate by keeping the update request with the highest side
effect counter for each leaf index. The "hints" in this case are the
number of times that each leaf index appears which we use in the circuit
with a new "assert_deduped_array" function.

NOTE: You'll notice that the expected orderings of update requests as
emitted by the tail is different now. That's because we sort by leaf
index, then side effect counter. I didn't see why we actually needed to
maintain the ordering across different slot. For example, what I've done
assumes it is okay if we have writes of (leaf, value, counter):
- (2, 2, 1)
- (1, 1, 2)
to have final output of
- (1, 1, 2)
- (2, 2, 1)

If that is not okay and for some reason we need to still have the writes
in their counter order after deduping then this will just make the
typescript initial sort slightly more complicated since we'll need to
arrange the slots such that they're sorted on side effect counter after
deduping.